pyproject.toml
[project]
name = "support-ops-env"
version = "0.1.0"
description = "SaaS Support Operations RL Environment"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "huggingface_hub",
    "openai",
    "requests",
    "python-dotenv",
    "websockets",
    "openenv-core",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server", "agent"]

[tool.uv]
package = true